Move creatable group types to /models
