+++ src/models/group/to-raw-creatable-group.ts
@@ -6,10 +6,10 @@
  * MIT license. See the LICENSE file for details.
  **************************************************************************/
 
-export * from './creatable-group';
-export * from './group';
-export * from './is-group';
-export * from './raw-creatable-group';
-export * from './raw-group';
-export * from './to-group';
-export * from './to-raw-creatable-group';
+import { CreatableGroup } from './creatable-group';
+import { RawCreatableGroup } from './raw-creatable-group';
+
+export const toRawCreatableGroup = (creatable: CreatableGroup): RawCreatableGroup => ({
+	Name: creatable.name,
+	Desc: creatable.description ?? '',
+});
